Mostly Completed Player Interaction System
All sequential events are now working in the test level. Player can pick up an item, carry it, inspect it, rotate it, zoom in and out on it, and proper UI appears and disappears when necessary to show user how to use the interaction system. All that needs to be done now is to enable the player to stop moving when picking up the item and being able to drop it after viewing. Or in other words, taking off the carrying part of the inspection process.