Finished functionality of new Settings Menu UI
Created new widgets for Controls, Audio, and Video UI for the Settings Menu in the Main Menu. Designed Blueprints for full functionality. Player can now see an updated version of the Settings Menu that is more interactive and aesthetically pleasing. They can click on the available options (Controls, Audio, and Video) and in Audio the player can manipulate the SFX, Master Volume, and Music Volume individually in real time. The player can also click on the back menu to go back to the Main Menu and repeat the process to their desire.